fix(dispatch): lift role-injection to lane-neutral skill_context module (PR-1313 r2)
The envelope/provider lanes imported _inject_skill_context from
subprocess_dispatch_internals.skill_injection, crossing into the subprocess
lane's internal package — VNX CI Profile A flagged dispatch_envelope.py for
new direct tmux/subprocess coupling. The guard is right and is not bypassed;
dispatch_envelope.py is not added to any exception list.

The worker prompt-context injection moves to scripts/lib/skill_context.py
(lane-neutral, top-level shared module). subprocess_dispatch_internals/
skill_injection.py is now a thin compat re-export, so the pre-existing users
(tmux_interactive_dispatch.py:527, dispatch_prepare.py:67, the subprocess
facade) keep working unchanged. dispatch_envelope._prepare and
provider_dispatch._enrich_instruction now import from the canonical module.

Part B (role_application.py) and the Part A wiring behavior are untouched.
Test patch targets moved to skill_context where they patch the moved
functions' internals. ruff clean on all changed source files.
