+++ scripts/lib/skill_context.py
@@ -1,0 +1,342 @@
+"""skill_context — lane-neutral worker prompt-context injection (canonical home).
+
+Shared, lane-neutral implementation of worker prompt-context assembly used by
+every dispatch lane without any lane importing from another lane's internal
+package. The functions here were moved from
+``subprocess_dispatch_internals/skill_injection.py``; that module is now a thin
+compat re-export of this one, so existing imports and test patches at the old
+path keep resolving.
+
+Layers:
+  - _inject_skill_context        — role context + intelligence + full assembly
+  - _inject_permission_profile   — permission-profile preamble
+  - _build_intelligence_section  — intelligence items (delegates to
+                                   intelligence_injection.fetch_intelligence_section)
+  - legacy 3-tier CLAUDE.md resolution + PromptAssembler path
+
+Project-root and state-dir lookups resolve via the ``subprocess_dispatch``
+facade namespace (deferred imports inside the functions) so test monkeypatches
+at the facade keep intercepting — see the per-function docstrings.
+"""
+
+from __future__ import annotations
+
+import logging
+from pathlib import Path
+
+logger = logging.getLogger(__name__)
+
+
+def _inject_permission_profile(terminal_id: str, role: str | None, instruction: str) -> str:
+    """Prepend permission preamble to instruction if a profile exists for role.
+
+    Resolves the terminal's expected role from terminal_assignments when role
+    is None.  Logs a warning on role/terminal mismatch.  Returns instruction
+    unchanged when no profile is found or worker_permissions cannot be loaded.
+    """
+    try:
+        from worker_permissions import (
+            load_permissions,
+            generate_permission_preamble,
+            validate_dispatch_permissions,
+        )
+    except ImportError:
+        logger.debug("_inject_permission_profile: worker_permissions not available, skipping")
+        return instruction
+
+    effective_role = _resolve_effective_role(terminal_id, role)
+    if not effective_role:
+        return instruction
+
+    warnings = validate_dispatch_permissions(
+        {"terminal": terminal_id, "role": effective_role}
+    )
+    for w in warnings:
+        logger.warning(w)
+
+    profile = load_permissions(effective_role)
+    if not profile.allowed_tools and not profile.denied_tools and not profile.bash_deny_patterns:
+        logger.debug("_inject_permission_profile: empty profile for role '%s', skipping preamble", effective_role)
+        return instruction
+
+    preamble = generate_permission_preamble(profile)
+    logger.info(
+        "Permission profile applied: terminal=%s role=%s allowed=%s denied=%s",
+        terminal_id, effective_role,
+        profile.allowed_tools,
+        profile.denied_tools,
+    )
+    return f"{preamble}\n---\n\n{instruction}"
+
+
+def _resolve_effective_role(terminal_id: str, role: str | None) -> str | None:
+    """Resolve effective role from terminal_assignments when role is None.
+
+    Uses the project-override-first path resolution from worker_permissions so
+    central-install and embedded installs both find the project's customised
+    worker_permissions.yaml rather than the immutable VNX_HOME template.
+    """
+    if role:
+        return role
+    try:
+        from worker_permissions import _resolve_permissions_yaml  # noqa: PLC0415
+        yaml_path = _resolve_permissions_yaml()
+        import yaml  # noqa: PLC0415
+        data = yaml.safe_load(yaml_path.read_text()) or {}
+        return data.get("terminal_assignments", {}).get(terminal_id)
+    except Exception as exc:
+        logger.debug("_inject_permission_profile: could not resolve role for %s: %s", terminal_id, exc)
+        return None
+
+
+def _build_intelligence_section(
+    dispatch_id: str,
+    role: str | None,
+    *,
+    dispatch_paths: "list[str] | None" = None,
+    instruction_text: str | None = None,
+    pr_id: str | None = None,
+) -> str:
+    """Return formatted intelligence items as markdown, or empty string (best-effort).
+
+    Delegates to intelligence_injection.fetch_intelligence_section so the same
+    logic serves all provider paths (codex/gemini/litellm/kimi) via
+    provider_dispatch.py.
+
+    _default_state_dir is fetched via subprocess_dispatch namespace so test
+    monkeypatches at the facade honour this call site unchanged.
+
+    dispatch_paths, instruction_text, pr_id are forwarded to selector.select() so
+    W5 item classes (adr_relevant, code_anchor, operator_memory, schema_section,
+    prior_round_finding) can fire in production dispatches.
+    """
+    try:
+        from intelligence_injection import fetch_intelligence_section  # noqa: PLC0415
+    except ImportError as exc:
+        logger.warning("intelligence injection failed (%s); proceeding without", exc)
+        return ""
+    # Look up _default_state_dir via subprocess_dispatch namespace so test
+    # monkeypatches at the facade ("subprocess_dispatch._default_state_dir")
+    # are honoured when this helper is called directly by tests.
+    import subprocess_dispatch as _sd
+    state_dir = _sd._default_state_dir()
+    return fetch_intelligence_section(
+        dispatch_id=dispatch_id,
+        role=role,
+        state_dir=state_dir,
+        pr_id=pr_id,
+        dispatch_paths=dispatch_paths,
+        instruction_text=instruction_text,
+    )
+
+
+def _inject_skill_context(
+    terminal_id: str,
+    instruction: str,
+    role: str | None = None,
+    dispatch_metadata: "dict | None" = None,
+) -> str:
+    """Compose layered user message context for headless dispatch.
+
+    Uses PromptAssembler (3-layer architecture) when available, with fallback
+    to the legacy 3-tier CLAUDE.md resolution for backward compatibility.
+
+    Layer architecture (PromptAssembler path):
+      Layer 1 — Base worker context (universal rules, report format)
+      Layer 2 — Role context (capabilities, permissions for the role)
+      Layer 3 — Dispatch payload (passed through as instruction)
+
+    Legacy fallback (3-tier CLAUDE.md resolution):
+      1. agents/{role}/CLAUDE.md        — project-level agent override
+      2. .claude/skills/{role}/CLAUDE.md — skill definition
+      3. .claude/terminals/{terminal}/CLAUDE.md — terminal fallback
+
+    Args:
+        terminal_id:       Terminal identifier (e.g. "T1").
+        instruction:       Raw dispatch instruction text.
+        role:              Agent role (e.g. "backend-developer").
+        dispatch_metadata: Optional metadata dict forwarded to PromptAssembler
+                           for L3 enrichments (dispatch_id, gate, pr, track, model,
+                           intelligence, historical).  Merged with terminal+role.
+
+    Returns the full pipe_input string ready for `claude -p`.
+    """
+    # Gather intelligence before assembling prompt (best-effort).  Looked up via
+    # subprocess_dispatch namespace so test patches at the facade are honoured.
+    import subprocess_dispatch as _sd
+    _dispatch_id = (dispatch_metadata or {}).get("dispatch_id") or ""
+    _dispatch_paths = (dispatch_metadata or {}).get("dispatch_paths") or []
+    _instruction_text = instruction
+    _pr_id = (dispatch_metadata or {}).get("pr_id") or (dispatch_metadata or {}).get("pr")
+    intelligence_section = _sd._build_intelligence_section(
+        _dispatch_id, role,
+        dispatch_paths=_dispatch_paths,
+        instruction_text=_instruction_text,
+        pr_id=_pr_id,
+    )
+
+    assembled = _try_prompt_assembler(
+        terminal_id, instruction, role, dispatch_metadata, intelligence_section,
+    )
+    if assembled is not None:
+        return assembled
+
+    return _legacy_claude_md_resolution(
+        terminal_id, instruction, role, intelligence_section,
+    )
+
+
+def _has_prompt_assembler_role(role: str | None) -> bool:
+    """True when PromptAssembler has a role prompt for *role*.
+
+    PromptAssembler's L2 reads ``scripts/lib/prompts/roles/<role>.md`` and falls
+    back to base_worker.md when the file is missing — so two distinct roles
+    WITHOUT a prompt file (e.g. quality-engineer vs system-architect) would
+    otherwise receive byte-identical context.
+    """
+    if not role:
+        return False
+    # This module lives in scripts/lib (one level above the old
+    # subprocess_dispatch_internals home), so the prompts dir is parent/…,
+    # not parent.parent/….
+    return (Path(__file__).resolve().parent / "prompts" / "roles" / f"{role}.md").exists()
+
+
+def _has_legacy_role_source(role: str | None) -> bool:
+    """True when the legacy 3-tier has a ROLE-SPECIFIC source (agents/ or skills/).
+
+    Terminal fallback is deliberately excluded: it is generic per terminal, not
+    role context. Only role-specific sources decide the routing below.
+    """
+    if not role:
+        return False
+    import subprocess_dispatch as _sd
+    root = _sd.Path(_sd.__file__).resolve().parents[2]
+    return (
+        (root / "agents" / role / "CLAUDE.md").exists()
+        or (root / ".claude" / "skills" / role / "CLAUDE.md").exists()
+    )
+
+
+def _try_prompt_assembler(
+    terminal_id: str,
+    instruction: str,
+    role: str | None,
+    dispatch_metadata: "dict | None",
+    intelligence_section: str,
+) -> str | None:
+    """Attempt PromptAssembler path; return assembled pipe_input or None on failure.
+
+    Dispatch-20260801-w10 routing fix: when the role has a project-level
+    agents/skills source but NO PromptAssembler role prompt, return None to route
+    through the legacy 3-tier resolution. PromptAssembler would otherwise
+    silently substitute base_worker.md as L2, so a dispatch with
+    ``role=quality-engineer`` got the exact same context as one with
+    ``role=system-architect``. Roles WITH a prompt file (backend-developer,
+    test-engineer, ...) keep using PromptAssembler unchanged.
+    """
+    if _has_prompt_assembler_role(role) is False and _has_legacy_role_source(role):
+        logger.info(
+            "_inject_skill_context: role=%s has no PromptAssembler prompt but has a "
+            "legacy role source — using legacy 3-tier so agents/{role}/CLAUDE.md reaches the worker",
+            role,
+        )
+        return None
+    try:
+        from prompt_assembler import PromptAssembler  # noqa: PLC0415
+        assembler = PromptAssembler()
+        meta = dict(dispatch_metadata or {})
+        meta.setdefault("role", role or "")
+        meta.setdefault("terminal", terminal_id)
+        if intelligence_section:
+            meta.setdefault("intelligence", intelligence_section)
+        assembled = assembler.assemble(
+            dispatch_metadata=meta,
+            instruction=instruction,
+        )
+        logger.info(
+            "_inject_skill_context: assembler path — role=%s L1=%d L2=%d L3=%d chars",
+            assembled.metadata.get("role"),
+            assembled.metadata.get("layer1_chars", 0),
+            assembled.metadata.get("layer2_chars", 0),
+            assembled.metadata.get("layer3_chars", 0),
+        )
+        return assembled.to_pipe_input()
+    except Exception as exc:
+        logger.warning(
+            "_inject_skill_context: PromptAssembler failed (%s) — falling back to legacy CLAUDE.md resolution",
+            exc,
+        )
+        return None
+
+
+def _legacy_claude_md_resolution(
+    terminal_id: str,
+    instruction: str,
+    role: str | None,
+    intelligence_section: str,
+) -> str:
+    """3-tier CLAUDE.md resolution fallback.
+
+    Resolves project root via the facade's ``__file__`` (same pattern as
+    ``_resolve_agent_cwd``) so tests that patch ``subprocess_dispatch.__file__``
+    can intercept the resolution (test_skill_context_injection).
+    """
+    import subprocess_dispatch as _sd
+    project_root = _sd.Path(_sd.__file__).resolve().parents[2]
+
+    candidates: list[Path] = []
+    if role:
+        candidates.append(project_root / "agents" / role / "CLAUDE.md")
+        candidates.append(project_root / ".claude" / "skills" / role / "CLAUDE.md")
+    candidates.append(project_root / ".claude" / "terminals" / terminal_id / "CLAUDE.md")
+
+    for path in candidates:
+        if path.exists():
+            context = path.read_text()
+            if intelligence_section:
+                return (
+                    f"{context}\n\n---\n\n"
+                    f"## Relevant Intelligence (from past dispatches)\n\n"
+                    f"{intelligence_section}\n"
+                    f"---\n\nDISPATCH INSTRUCTION:\n\n{instruction}"
+                )
+            return f"{context}\n\n---\n\nDISPATCH INSTRUCTION:\n\n{instruction}"
+
+    if intelligence_section:
+        return (
+            f"## Relevant Intelligence (from past dispatches)\n\n"
+            f"{intelligence_section}\n"
+            f"---\n\nDISPATCH INSTRUCTION:\n\n{instruction}"
+        )
+    return instruction
+
+
+def _resolve_agent_cwd(role: str | None) -> Path | None:
+    """Return agents/{role}/ as Path if the directory exists, else None.
+
+    Resolves project root via the facade's ``__file__`` and ``Path`` so tests
+    that patch ``subprocess_dispatch.Path`` can intercept the resolution
+    (test_subprocess_dispatch_f34.TestDeliverCwdPropagation).
+    """
+    if not role:
+        return None
+    import subprocess_dispatch as _sd
+    candidate = _sd.Path(_sd.__file__).resolve().parents[2] / "agents" / role
+    return candidate if candidate.is_dir() else None
+
+
+def _load_agent_profile(config_path: Path) -> str:
+    """Load governance_profile from agent config.yaml.
+
+    Uses a simple line-scan so no yaml dependency is required.
+    Returns 'default' when the key is absent or the file cannot be read.
+    """
+    try:
+        for line in config_path.read_text().splitlines():
+            line = line.strip()
+            if line.startswith("governance_profile:"):
+                return line.split(":", 1)[1].strip()
+    except Exception as _exc:
+        logger.warning("Failed to read agent config %s: %s", config_path, _exc)
+    return "default"
